refactor rx retrieve for better maintainability
Hi!

I wanted to give you feedback on your `retrieve()` function but thought it was the easiest by providing you with this PR. 

Note; i'm usually using Rxjs 4.x and i know that they have not implemented all operators and renamed a few in 5.x. This PR might not 100% run due to this.